[refactor] #57 likeService Redis 의존성 분리
