+++ src/main/java/com/numble/team3/like/domain/LikeUtils.java
@@ -1,0 +1,11 @@
+package com.numble.team3.like.domain;
+
+import com.numble.team3.like.application.response.GetVideoRankDto;
+import java.util.List;
+
+public interface LikeUtils {
+
+  void processChangeDayRanking(String standard, List<GetVideoRankDto> ranking);
+
+  List<GetVideoRankDto> getDayRanking(String standard);
+}
